Use `@abstractmethod` for abstract method
Summary:
For some reason we were incorrectly marking the `Typeshed` class's
methods as `abc.abstractclassmethod`. This causes problems in cI
because of... I'm not sure, but basically a Pyre bug plus version skew
probably; it was a type error internally and an unused suppression in
github CI, I filed T189622394 to investigate.

But I also realized that `abstractclassmethod` is actually wrong anyway,
these are normal methods. So just using `abc.abstractmethod` fixes both
the CI issues and some actually-wrong code.

Reviewed By: grievejia

Differential Revision: D57531832

fbshipit-source-id: 659a52941d680b5a8ae20e876754336be6f37210