Support --no-logger to override logger configuration from CLI
Summary:
To create a Configuration, Pyre creates and merges PartialConfiguration objects
in precedence order. First .pyre_configuration, a local/codenav configuration,
and finally command line arguments.

However, it's not possible to *unset* a configuration. For example, if the
.pyre_configuration file specifies a logger, we can't override that to be `None`
with the simple merge operation.

This diff adds support for --no-logger in a somewhat hacky way. We add a field
to CommandArguments which is populated by the CLI arguments. This is not
converted into PartialConfiguration, but rather handled specially at the end.

I'm open to suggestions for nicer ways to do this. I tried a few alternatives
already, but they were much more invasive and didn't feel much cleaner to me. I
think the right thing to do here is to eliminate `from_command_arguments`
entirely and code CLI overrides separately from configuration overrides, but
this involved way too many changes for my appetite.

Reviewed By: grievejia

Differential Revision: D57591850

fbshipit-source-id: ebab84115a49339b66007cef7edbd31ef4850d75