Refactor the apply function to be at the top level in attributeResolution
Summary:
We are going to expose all dataclass functionality to the typechecker. To do so, we must separate that functionality from attributeResolution.
The apply function is related to attributeResolution, so we must move it to the top level of attributeResolution and separate it from the rest of the dataclass functionality (which we will move to a different file in the next diff)

Reviewed By: stroxler

Differential Revision: D57584806

fbshipit-source-id: d96496ef73eba54e65b42dd902dec59491b43b56