Delete the multi-source issue post-processing step
Summary:
For context, the multi-source issues are hard to triage in the UI, mainly
because the subtraces are confusing.

This diff stack attempts to drop the support of allowing a user to specify main
/ secondary flows. I think it is easier to add the support (of specifying main
/ secondary flows) back if later we decide this is indeed useful, than to keep
the support while doing various changes related with multi-source rules.
Later we will evaluate that, after dropping the support, if the multi-source
issues are still easy to triage. If so, then we keep it as is. Otherwise, we
might want to restore the support at least for a few situations.

This diff
- removes the multi-source issue post-processing step, since that is
used to allow a user to specify main / secondary flow,
- deletes the unused functions from module `Issue` that are related with
  multi-source issue post-processing, and
- updates `compute_triggered_flows` so that it no longer generates two issues
  for a multi-source rule. It now only generates an issue for the flow that
  ends up with triggered sinks (i.e., the flows that are discovered later by
  the analysis than the initially discovered flows).
  Since the post-processing step removes the secondary issues and this step is
  now removed, we should only generate one issue per multi-source rule. For
  simplicity and to align with other tools, I decided to restore the original
  behavior that reports the flows that are discovered later as the issues.
  This decision makes the specifications of main / secondary flows useless,
  which could be confusing to the users but this is an intermediate state that
  we will soon walk out of. We will either remove the parsing logic or
  restore the support for a limited situations.

The changes in `*.models` files are due to
- ignoring the main / flow specifications -- some secondary flows become the
  main flows (and vice versa),
- removing the multi-source issue handles, and
- removing the subtraces that show the flows for the partial sinks. This will
  be added back in the follow-up diffs.

The follow-up diffs will add back the subtraces that are used to show the flows
that end up with the partial sinks (i.e., the initially discovered flows).

Reviewed By: alexkassil

Differential Revision: D57414010

fbshipit-source-id: c8f913a679cf43c67e2bb30ba445f61d270636d8