Move the common argument types for analyzing joined strings into `CallModel`
Summary:
The arguments of function `analyze_joined_string` between the forward and
backward analysis should have the same types. This not only reflects the fact
that the function has the dual logic between the forward and backward analysis,
but also helps the developer to sanity check that the dual logic is done in
both places.

To achieve the above goal, this diff moves the common argument types into
`CallModel`. The next diff will refactor `analyze_joined_string` in the
backward analysis, to use the common argument types.

This diff also moves function `implicit_string_literal_sources` into
`CallModel`. Although this function is only used by the forward analysis (and
not by the backward analysis), the backward analysis will have a dual function
that will also be moved into `CallModel`, so that the duality between the two
functions are explicit.

Reviewed By: arthaud

Differential Revision: D57689979

fbshipit-source-id: 59bef3db5918709f8188b7ad8a96c01e3f6da6d4