allow trailing named arguments to match already-consumed parameters
Summary:
In attributeResolution, arguments & parameters are matched in order and a starred argument greedily consumes all possible positional parameters.

This is done without looking at how many items are in the expanded tuple (which leads to some other known issues), but sometimes we actually don't know the arity so we do have to consume everything just to be safe.

If *args is followed by a named argument, odds are the parameter has already been consumed and we'll get an unexpected keyword error. I think this behavior errs too far on the safe side and can lead to friction - see the new type errors from the Pyre CI job in D57089916

In this diff, I allow the named argument to be associated with a parameter that has already been consumed, if that parameter has not been matched with another positional or explicitly named argument.

Reviewed By: stroxler

Differential Revision: D57675482

fbshipit-source-id: 8c69ce400e5adda3da9a58d118bbe89266e936c5