Fix upgrade Configuration find nested configurations to use filesystem
Summary: When this logic was created, I did a simple `glob.glob()` search to query for files that should be included by a configuration. This logic is correct, but it's much more efficient to do through `hg` if possible. This uses the `filesystem.Filesystem` logic that makes that functionality much faster and easier to do.

Reviewed By: MaggieMoss

Differential Revision: D57678576

fbshipit-source-id: 636100ffa5434980f6a93a3cdcad38a578029d7c