Move exclude filter to get files logic
Summary:
In finding files covered by a configuration, we would attempt to filter out files that are covered by nested configurations by determining which files are present in a nested configuration and excluding them from "this configuration's" included files. We would later do the work to filter out files that should be excluded or ignored when doing work around the mode header, but that's incorrect.

By doing that logic, a file that's excluded might not be found by a parent configuration, even if it should be owning it, because we would consider it 'owned' by the nearest .pyre_configuration.local.

This diff moves the exclusion logic to the file filtering, so that we can properly determine which files belong to which configurations.

This pattern should be extremely uncommon, but might as well fix it now.

Reviewed By: MaggieMoss

Differential Revision: D57678626

fbshipit-source-id: a05a2a61ef7f4bed1d80293199fca776fa3f8408