Infer taint propagation from arguments to self
Summary:
We currently only infer taint propagations (i.e tito) from arguments to self for constructors, `__setitem__` and property setters.
This diff enables it for all methods marked with the `InferSelfTito` mode or when `--infer-self-tito` is provided via the command line.
In general, this allows to find more results at the cost of performance and false positives, hence why we add it behind a flag for now.

Reviewed By: alexkassil

Differential Revision: D57724344

fbshipit-source-id: 99a598d6a00239a72e52db4741493b910c0fb25a