Add a testcase to show that assigning a Callable variable to a dataclass field makes it a BoundMethod.
Summary:
The code in the given unittest fails with a Too many arguments type error.  Looking at the testcase, the problem is coming from this line matcher: Callable[[str], bool] = dataclasses.field(). In this line, matcher is a BoundMethod. Yet, if we remove the field assignment, matcher is a Callabe.

Adding this diff so we can figure out if this is normal behavior. The fact that matcher is a BoundMethod is causing it not to be a callable which is causing the type error.

Reviewed By: stroxler

Differential Revision: D57797359

fbshipit-source-id: d09a389666a47f1bf7c5e8d1ba9f43dae67b17ca