Generalize frozen checks to handle metaclasses and custom decorators
Summary:
In the previous diff of this stack, we refactored the dataclassOptions API and exposed it to the typechecker for the purpose of generalizing frozen checks.

In this diff, we utilize this API. Specifically, we handle the vanilla decorator case but we also handle the custom metaclass transform case, which is shown in our new unittest.

Reviewed By: stroxler

Differential Revision: D57789043

fbshipit-source-id: aad97164883c9c4eccb6f1cb7621b593ec629066