Infer taint propagations between arguments
Summary:
We currently do NOT infer propagations between arguments. For instance, if a function mutates an argument with the value from another argument, we won't find the flow.

For instance:
```
def list_append(l: List[str], v: str):
    l.append(v) # tito v -> l[*]
```

This diff allows infering those propagations when `--infer-argument-tito` is provided or when the function is marked with the mode `InferArgumentTito`.

Reviewed By: tianhan0

Differential Revision: D57739123

fbshipit-source-id: 5dcefbf42b1d3aeff44748d352e6519e41fe433a