Add caching to configurationless migration (#866)
Summary:

This diff adds functionality around caching operations determining which files are included in a given .pyre_configuration.local.

Determining which files are included can be expensive. We need to perform operations to determine the following:
* What are all files that can be included in this configuration? We ask `hg` (or perform a glob) for this
* What are all files that should be excluded from this configuration?
* What are all files included in nested configurations where those configurations should have priority over this one? We perform the first step in this list once for each nested configuration.

The steps above are performed once for each file we're linting, so determining which files should be included is an expensive process. We attempt to mitigate it by caching a lot of these operations.

Reviewed By: MaggieMoss

Differential Revision: D57681433