Parse using the same partial sink kind in multiple rules, and using multiple partial sink kinds in a single rule
Summary:
We would like to use the same partial sink kind in multiple rules, which is not
currently allowed. In addition, we would also want to use multiple partial sink
kinds, in a single multi-source rule.

This diff enables the parsing of the above features. The follow-ups will enable
the taint analysis to handle them.

Implementation-wise, since `PartialSinkLabelsMap` is a map from partial sink
kinds to their main and secondary labels, it does not support a partial sink
kind to have more than the main and secondary labels. This could happen when a
partial sink kind is associated with some labels in one rule but some other
lables in another rule.
To solve this problem, this diff simply lets a partial sink kind to map to a
list of associated labels, without restricting the associated labels to be only
main and secondary. One can essentially think of a partial sink kind together
with a label as a unique partial sink kind.

Reviewed By: arthaud

Differential Revision: D57507470

fbshipit-source-id: c0a8ba47ed5af82f0c25a992b5810e62a1e7af96