Add caching to configurationless migration (#866)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/866

This diff adds functionality around caching operations determining which files are included in a given .pyre_configuration.local.

Determining which files are included can be expensive. We need to perform operations to determine the following:
* What are all files that can be included in this configuration? We ask `hg` (or perform a glob) for this
* What are all files that should be excluded from this configuration?
* What are all files included in nested configurations where those configurations should have priority over this one? We perform the first step in this list once for each nested configuration.

The steps above are performed once for each file we're linting, so determining which files should be included is an expensive process. We attempt to mitigate it by caching a lot of these operations.

Reviewed By: MaggieMoss

Differential Revision: D57681433

fbshipit-source-id: 5f860c4c4c9f1ba5b1594a753d7b6455a60bc251