Delete shape types magic from builtins.pyi patches
Summary:
As with the use of `pyre_extensions.ReadOnly` in stdlib stubs, it isn't really
okay to be making use of `Add` / `Multiply` / `Divide` in stdlib stubs because
this will actually introduce any expressions in many projects (any project not
explicitly depending on `pyre_extensions`.

These patches only existed in order to support fancy syntax for shape types;
since we are expecting to cut shape type support anyway for the moment (in
its existing form it never found a user) I'm not trying to make a
a workaround, we can just get rid of the patches.

Reviewed By: connernilsen

Differential Revision: D57930998

fbshipit-source-id: 2369ce4c4d5fea175b17ff1244119adafe6cdcc8