Support typing._PyreReadOnly_ as ReadOnly
Summary:
We've known about a problem for over a hear where many projects wind up with Any expressions. The problem occurs if the project:

(a) does not use `pyre_extensions` and
(b) makes any use of builtin types that could potentially be read-only, such as subscripting dicts

The problem is that Pyre will hard-choke (meaning give up with no attempt to fall back, producing `type.Top` as the type and treating this like `Any` for most purposes) if it encounters unknown types in a signature. And we've patched the standard library to use `pyre_extensions.ReadOnly`!

One workaround is to just inject a made-up, private read-only generic into `typing` and rely on that in stdlib.

Since this workaround also naturally jives with resolving the name collision relative to PEP 705's type-dict ReadOnly, I want to just go ahead and push forward with it.

This diff modifies Pyre iteself to handle `typing._PyreReadOnly_`; I'll actually make the change to the stubs in a separate diff.

Reviewed By: connernilsen

Differential Revision: D57927336

fbshipit-source-id: 51b52a202409a45d9d72d0db8e295d31cdba29f6