Add unittest to show that we are incorrectly generating an error when calling constructor of dataclass with descriptor
Summary:
In an effort to improve conformance, we will investigate the remaining dataclass conformance failures and add unittests to cover them. Each yellow box will have about 2-4 items, and we will add a unittest to cover all of them, such that if the unittest passes, the box will turn green.

In this diff, we are considering this conformance test https://github.com/python/typing/blob/main/conformance/tests/dataclasses_descriptors.py

We also have a linked task detailing it.

Reviewed By: grievejia

Differential Revision: D57943302

fbshipit-source-id: 6bedcaaa917f34ca8d3d4191662b53d629e15579