Add a unittest that covers all dataclass_usage failures in the conformance tests
Summary:
In an effort to improve conformance, we will investigate the remaining dataclass conformance failures and add unittests to cover them. Each yellow box will have about 2-4 items, and we will add a unittest to cover all of them, such that if the unittest passes, the box will turn green.

In this diff, we are covering the dataclass_usage test:https://github.com/python/typing/blob/main/conformance/tests/dataclasses_usage.py

We also have a linked task detailing it.

Reviewed By: grievejia

Differential Revision: D57939097

fbshipit-source-id: 0e4231274a1ae4c00ef512a9417b7df34e258910