Add a unittest to show that we do not report when dataclasses are not compatable with hashable protocol
Summary:
In an effort to improve conformance, we will investigate the remaining dataclass conformance failures and add unittests to cover them. Each yellow box will have about 2-4 items, and we will add a unittest to cover all of them, such that if the unittest passes, the box will turn green.

In this diff, we are covering the dataclasses_hash https://github.com/python/typing/blob/main/conformance/tests/dataclasses_hash.py

We also have a linked task detailing it.

Reviewed By: grievejia

Differential Revision: D57940158

fbshipit-source-id: 4a4db0cc2d446d9da9f9f1ef490a6b93a82da723