Unittest to show that we do not propagate @dataclass_transform when used with @overload to the definition.
Summary:
In an effort to improve conformance, we will investigate the remaining dataclass conformance failures and add unittests to cover them. Each yellow box will have about 2-4 items, and we will add a unittest to cover all of them, such that if the unittest passes, the box will turn green.

In this diff, we are considering this conformance test https://github.com/python/typing/blob/main/conformance/tests/dataclasses_transform_func.py

We also have a linked task detailing it.

Reviewed By: grievejia

Differential Revision: D57945037

fbshipit-source-id: 6e4ada4864298d63bbca9407309826ac9dd82b89