Add an option to set the scheduling policy of map reduce operations [client 3/3]
Summary:
# Problem
We are running Pysa on bigger machines (more CPU and memory), and are getting analysis times similar to running on smaller machines.
We noticed a very low CPU usage, which is most likely due to the way we chunk our jobs across workers with map-reduce operations.
We need to tune the chunk size and/or chunk count for each map reduce operation depending on the machine type.

# Solution
This stack of diff makes the scheduler policy (i.e chunk size/chunk count/etc.) for each map reduce operation configurable by the user. This will make it easy to tune the parameters and experiment to find the best parameters.

The user can provide a "scheduler policies configuration file" using the command line option `--scheduler-policies <path>`.

The file will be a JSON object where keys are schedule identifiers (e.g, type_check, taint_fixpoint, etc.) and the value is a scheduler policy of the form:
```
SchedulerPolicy :=
  | { "kind": "fixed_chunk_size",
      "minimum_chunk_size": <int>,
      "minimum_chunks_per_worker": <int>,
      "preferred_chunk_size": <int>
    }
  | { "kind": "fixed_chunk_count",
      "minimum_chunks_per_worker": <int>,
      "minimum_chunk_size": <int>,
      "preferred_chunks_per_worker": <int>
     }
```

In the future, we could have predetermined profiles for specific machines, e.g using `--tier t10` or `--tier t6f` options.

# This diff
This diff implements the option on the client side, handling parsing of the json file and passing it to the ocaml binary.

Reviewed By: tianhan0

Differential Revision: D58019294

fbshipit-source-id: e5f6d9683bc5d92c5109036c983e10042c0b4617