Add an option to specify the scheduling policy of map reduce operations [ocaml 2/3]
Summary:
# Problem
We are running Pysa on bigger machines (more CPU and memory), and are getting analysis times similar to running on smaller machines.
We noticed a very low CPU usage, which is most likely due to the way we chunk our jobs across workers with map-reduce operations.
We need to tune the chunk size and/or chunk count for each map reduce operation depending on the machine type.

# Solution
This stack of diff makes the scheduler policy (i.e chunk size/chunk count/etc.) for each map reduce operation configurable by the user. This will make it easy to tune the parameters and experiment to find the best parameters.

The user can provide a "scheduler policies configuration file" using the command line option `--scheduler-policies <path>`.

The file will be a JSON object where keys are schedule identifiers (e.g, type_check, taint_fixpoint, etc.) and the value is a scheduler policy of the form:
```
SchedulerPolicy :=
  | { "kind": "fixed_chunk_size",
      "minimum_chunk_size": <int>,
      "minimum_chunks_per_worker": <int>,
      "preferred_chunk_size": <int>
    }
  | { "kind": "fixed_chunk_count",
      "minimum_chunks_per_worker": <int>,
      "minimum_chunk_size": <int>,
      "preferred_chunks_per_worker": <int>
     }
```

In the future, we could have predetermined profiles for specific machines, e.g using `--tier t10` or `--tier t6f` options.

# This diff
This diff updates most call sites of `Scheduler.map_reduce` to use a user-provided scheduler policy.
Note that some map-reduce calls from type checking are not covered, since this is not a priority AFAIK for pyre to handle this.

Reviewed By: grievejia

Differential Revision: D58019293

fbshipit-source-id: 9657e99ee94f8f0b30ed1ce822860bad11bc5ce2