Add an option to specify the scheduling policy of map reduce operations [ocaml 1/3]
Summary:
# Problem
We are running Pysa on bigger machines (more CPU and memory), and are getting analysis times similar to running on smaller machines.
We noticed a very low CPU usage, which is most likely due to the way we chunk our jobs across workers with map-reduce operations.
We need to tune the chunk size and/or chunk count for each map reduce operation depending on the machine type.

# Solution
This stack of diff makes the scheduler policy (i.e chunk size/chunk count/etc.) for each map reduce operation configurable by the user. This will make it easy to tune the parameters and experiment to find the best parameters.

The user can provide a "scheduler policies configuration file" using the command line option `--scheduler-policies <path>`.

The file will be a JSON object where keys are schedule identifiers (e.g, type_check, taint_fixpoint, etc.) and the value is a scheduler policy of the form:
```
SchedulerPolicy :=
  | { "kind": "fixed_chunk_size",
      "minimum_chunk_size": <int>,
      "minimum_chunks_per_worker": <int>,
      "preferred_chunk_size": <int>
    }
  | { "kind": "fixed_chunk_count",
      "minimum_chunks_per_worker": <int>,
      "minimum_chunk_size": <int>,
      "preferred_chunks_per_worker": <int>
     }
```

In the future, we could have predetermined profiles for specific machines, e.g using `--tier t10` or `--tier t6f` options.

# This diff
This diff implements the parsing of the option in OCaml.
The scheduler policies are only used for the taint fixpoint for now. See follow-up diffs.

Reviewed By: grievejia

Differential Revision: D58019295

fbshipit-source-id: e96ec4f9b5281e663be38a45d5355f5f6883b90f