Add unsafe_hash option to dataclasses
Summary:
In dataclasses, when unsafe_hash is set to True, we must enable __hash__.

In this diff, we enable this option so we can pick it up from the dataclass and check its value.

In the next diffs, we will use it to determine if hashing is allowed for a given dataclass.

Reviewed By: grievejia

Differential Revision: D57997961

fbshipit-source-id: 824da6c0c0e5fe33980f64de4adf1f6087884b0c