Delete shape types magic from builtins.pyi patches
Summary:
As with the use of `pyre_extensions.ReadOnly` in stdlib stubs, it isn't really
okay to be making use of `Add` / `Multiply` / `Divide` in stdlib stubs because
this will actually introduce any expressions in many projects (any project not
explicitly depending on `pyre_extensions`.

These patches only existed in order to support fancy syntax for shape types;
since we are expecting to cut shape type support anyway for the moment (in
its existing form it never found a user) I'm not trying to make a
a workaround, we can just get rid of the patches.

**Auto-update conformance results**

```
cd ~/fbsource/fbcode/tools/pyre/facebook/conformance/
buck2 run :conformance_output_script -- ./third_party
```

**Second try: redo of D57930998 backed out in D58020428**

Now that D58088396 and D58088397 have landed, this should no longer be risky - per target type checking is using an xplat copy of typeshed, not the development version.

Reviewed By: grievejia

Differential Revision: D58137209

fbshipit-source-id: ad8219712a1e11ca42ef0a6f67f10599c33a55aa