Remove monomial / polynomial types
Summary:
There has been no uptake on Pyre shape types in their current form despite
them existing for 2 years and a PyCon talk.

This is something we still might be interested in eventually, but

(a) The general feeling on the team is that using literal types is probably not
the right way to go, it is confusing and also syntactically pretty heavy.

(b) Before we can really push for shape types, we need to make sure Pyre's fundamentals
are solid, which means a lot of refactoring. The shape type code is complex and is a lot
of currently-unused baggage to carry across refactors.

I'm a little sad to remove this, but if we are able to revive the project someday,
then we can come back to this commit in order to find the old code for supporting
polynomial operations + broadcast, which could give us a big jump-start on any new
attempt at shape types.

Reviewed By: samwgoldman, connernilsen

Differential Revision: D57975538

fbshipit-source-id: 296db2134c42172405761077047d08c695d1eaed