Distinguish triggered sinks by the sources that trigger them
Summary:
Since now a partial sink kind can be used in multiple rules, it is possible
that different multi-source flows result in creating the same triggered sink.
Hence, we need to distinguish which flows result in creating a triggered sink,
so that we can create different issues according to the flows that result in
creating triggered sinks. Otherwise we'd have to create all possible issues (to
be safe), which means potential false positives.

Implementation-wise, the main changes are:
1. In `sinks.mli`, I create type `triggered_sink` to distinguish which source
results in creating a triggered sink. In addition, this can be used in the
result of parsing a rule, so that even if two rules use the same partial sink
kind, we can still distinguish their flows based on the triggered sinks -- they
would be different because their triggering sources are different.
By contrast, type `partial_sink` does not need a `triggering_source` field,
because a partial sink can be used in multiple rules and hence we implicitly
allow a partial sink to match against any source.
2. In `issue.ml`, I updated the value type of `TriggeredSinkHashMap.t` into a
map from sources to extra traces. Such information will be needed when
converting partial sinks into triggered sinks.
3. In `kindCoverage.ml`, when computing the kind coverage, we convert partial
sinks to all possible triggered sinks that the partial sink can be converted
to, which represents the "coverage". We cover any triggered sink that can be
converted from this partial sink.

Reviewed By: arthaud

Differential Revision: D58027571

fbshipit-source-id: f9aa24ee6f1fc645ae3765fc7637ebf202f84e60