Remove field `label` from `Sinks.partial_sink`
Summary:
So that in the next diff we can transition from using labels in
partial sink kinds into using only partial sink kinds in the rules and models.
This diff drops the backend support, and the follow-ups will drop the frontend
(parsing) support.

Implementation-wise, the main change is to rename `PartialSinkLabelsMap` into
`RegisteredPartialSinks`, because
1. we no longer have a concept of "labels", and
2. `PartialSinkLabelsMap` was used to check whether a partial sink kind has
been "registered" / "defined". A partial sink kind is defined if it is used in
some rule.

Reviewed By: arthaud

Differential Revision: D58089361

fbshipit-source-id: a12ce271a686b2fd24c44b606eb816596253f709