Create a Constructors module in `type.ml`
Summary:
I'm trying to make the `Type` module easier to understand
and change. The biggest improvement we could make is to get rid of all
the `Record` modules, which have pretty unreadable names, and use recursive
modules instead.

But I'll need to do quite a bit of pushing code around first, because
currently `type.ml` is freely mixing top-level helper functions with the
modules that "shadow" record modules like `Type.Callable`.

A first step toward figuring out whether this is even feasible is to start
grouping together some of these top-level functions so they are namespaced
and can be moved around, which also makes the dependency structure
clearer.

This diff pulls all the "smart constructor" helpers into a `Constructors`
module. I `include Constructors` at the bottom but inside of the body of
`type.ml` all calls to the constructors are namespaced either explicitly
or with a local `let open Constructors in` binding.

Reviewed By: grievejia

Differential Revision: D57991589

fbshipit-source-id: a4eaf2509f87413a511a41c7e24c082f66c97796