Move `create_readonly` into `Constructors`
Summary:
This diff continues rearranging `type.ml` so that I can turn it into
recursive modules and get rid of the `Record` module. Here I move
`create_readonly`, which was kind of floating all on its own, to live
with all the other smart constructors in the new `Constructors` module.

Reviewed By: yangdanny97

Differential Revision: D58171616

fbshipit-source-id: c077a0d962ef47547a1e4282fee01310c284bbae