add typechecking for except clauses
Summary:
Based on the documentation for exception groups and try/catch, annotated `except` clauses should only catch `BaseException` and not `BaseExceptionGroup` even though the latter extends the former.

This diff updates the CFG to expose Try statements to the typechecker so we can check the exception handlers, and updates uninitializedLocalCheck to explicitly ignore them.

When resolving the type annotation in the exception handler, I first try to parse it as a type name, and then if it fails I try to parse as an type variable expression to get some `type[T]` and extract the `T`.

This should align with the behavior of mypy and pyright since both of them allow type variables in this position.

Reviewed By: stroxler

Differential Revision: D58018874

fbshipit-source-id: 969aff5240b2a131e2129aaf6b2b592230914f55