Make a `Type.Extractions` submodule, move more predicates
Summary:
There are a number of `[a-z_]+_value` helper functions in `type.ml`
that return some inner type when an outer type matches a certain pattern.

These have to be included in a mutually recursive submodule chain because parts
of Variable depend on predicates that in turn depend on these value extractors,
and in any case it's nice to group them together so they are easier to find.

After making that change, I was also able to move more predicates into
the `Predicates` module, which unblocks moving `Variable` and `TypedDictionary`
up (I'll do that in the next diff).

Reviewed By: yangdanny97

Differential Revision: D58171601

fbshipit-source-id: fb7b8c32065d13bc53291fade9fa165a191050d9