Move modules around
Summary:
I want to create a `Predicates` module which will take care of the bulk of the
remaining top-level helper functions that need to be moved into modules in order
to flip `type.ml` to using recursive modules.

To do this, I need to move `Constructors` (which is used by `VisitWithTransform`
and `Visitors` up to near the top, since some of the predicates depend on visitors.

Reviewed By: yangdanny97

Differential Revision: D58171611

fbshipit-source-id: 59271d002cd8d0e81105caeb6062e459facde453