Move `Variable` and `TypedDictionary` up
Summary:
This commit moves the `Variable` and `TypedDictionary` type-variant
modules up toward the top of `Type.ml` (actually in the diff view it
maybe looks like I moved all the intervening code "down" but that's
the same thing).

This moves us one big step closer to being able to use recursive submodules
and get rid of the `Record` module.

Reviewed By: yangdanny97

Differential Revision: D58171605

fbshipit-source-id: 81bbf2f0e07385946e634ac738503217c37189f7