make `Type.Alias` module
Summary:
The `Type.Variable` module depends on the `alias` type, so we need
to make a module for alias so that it can be included in our recursive
submodules. It's probably better for consistency and ergonomics anyway
to have a `Type.Alias` module rather than let it float at the top level
of `Type`, which is a huge file.

Reviewed By: yangdanny97

Differential Revision: D58171600

fbshipit-source-id: 8b6d7131b0a24671e198293b6e186df3422ad91c