Create Type.Predicates module
Summary:
Move lots of predicates that take a type and return a boolean into a common module.

The end goal here is again to simplify `Type.ml` by eliminating the `Record`
module and making all of the variant types non-generic, although I think that
there is some small readability benefit even to just the namespacing and
grouping-together happening in these refactors.

Reviewed By: yangdanny97

Differential Revision: D58171606

fbshipit-source-id: f0164f2fa2b958b82a6f618ff56e44031481ed6a