Make a `Type.ToExpression` module, split logic a bit
Summary:
In order to convert `type.ml` to recursive submodules, we do have to include
the `expression` function in the recursion at least at first, because the
`TypedDictionary` module has functions that depend on it.

Since I am putting it in a stand-alone module, I figured I'd take the opportunity
to also split out some of the nested functions - I can see why we had them nested
because we don't want more top-level functions in `type.ml`, but once we have
a submodule that's nicely namespaced we can break things up and it's definitely
a little easier to skim.

Reviewed By: yangdanny97

Differential Revision: D58171599

fbshipit-source-id: bb8db9600bf2166b45ba4fa8fd6933f0a2fba428