Add a test showing Pyre symbol collection fails on nested targets
Summary:
I was noticing that `UnannotatedGlobal.t` assumes you can unambigously indicate
an assignment using a single tuple index, which suggested to me that the logic isn't
properly recursive. Looks like we just ignore assignment statements that don't match
our expectations about structure 100% (even the outermost names are never bound).

Interestingly I've never seen an actual bug report on this, so it might be low priority,
but it's still a limitation we ought to track. If a library ever *did* use this, we
would throw no-such-name errors all over the place in user code which would be problematic.

Reviewed By: yangdanny97

Differential Revision: D58241708

fbshipit-source-id: bec564ec909cd2c6c3c2a12e1fc70d82c1a24422