delete empty contributing docs
Summary:
both contributing.md and fb/contributing.md are blank, so I removed them.

the sidebar still has an entry for Contributing, but it links to a different place now.

in the external site, it links to CONTRIBUTING.md in the repo: https://github.com/facebook/pyre-check/blob/main/CONTRIBUTING.md

in the internal site, it links to the "Development - Getting Started" doc

Reviewed By: samwgoldman

Differential Revision: D58299990

fbshipit-source-id: f06137bb3bb290c2a672bd761a03d5a25511b375