rename forward_assign to forward_value_assignment and move it outside of forward_assignment
Summary:
Changes:

This diff moves the forward_assign helper function outside of forward_assignment to allow for reuse, and updates some variable/parameter names to be clearer.
- explicit -> has_explicit_annotation
- guide -> guide_annotation_type
- original_annotation -> unwrapped_annotation_type (we get this by resolving the annotation expression to a type, and unwrap Final/Initvar/ClassVar)

semantically this should be a no-op

Motivation:

The main reason I wanted to lift this particular piece of code out is that it contains the bulk of the logic for assignments, but crucially it also runs after we resolve the type for the RHS of the assignment.

This means that we can do whatever special typechecking/processing we need for the RHS of augmented assignments before calling this function.

Reviewed By: samwgoldman

Differential Revision: D58291288

fbshipit-source-id: 0432f6d13a501e466c058de8933318b1e997ef75