Rename `Refinement` to `TypeInfo`
Summary:
The name `refinement` suggests that this module is specifically
about type refinement, but that's not really the case - it's related,
but this module is actually where we store type context information.

The reason `Refinement` sort of made sense is that type refinement
requires putting the refined types *into* the type context, but that
doesn't make the type context "about refinement"; in reality, the
local types (in the absense of refinement, just the normal types
coming from type inference!) are also stored here.

After discussing with Jia, we think that the not-so-descriptive name
`TypeInfo` is probably best for now; it's better to have a not very
descriptive name so that users look at the code and comments than
to have a poorly chosen more-descriptive name that could be confusing.

----

Note that I also intend to move `Annotation.t`, which is also very poorly
named (it isn't an annotation at all, it's a type plus some mutability info
needed for us to know how refinement should behave) into this module, which
will get rid of one of the worst-named types in Pyre and will potentially
free up the `Annotation.t` name for use if we ever want to model annotations
using an IR that is distinct from the "fully resolved" types in `Type.t`.

That will be a little trickier becuase I have to break a dependency cycle
where TypeInfo depends on GlobalResolution depends on Annotation, but I'm
hoping to tackle it in the next week or two.

Reviewed By: samwgoldman

Differential Revision: D58306709

fbshipit-source-id: f5aba626c56344141dd0fab9d3673624c2875348