Set the minimum chunk size to 1
Summary:
# Problem
We have noticed very low CPU usage on big machines.
After investigation, it seems due to scheduler policies having a `minimum_chunk_size` of 100.
When the number of task is low, say 1000, and number of workers is high (say 80), using the preferred chunk size would lead to a number of tasks per chunk below that threshold.
This ultimately leads the scheduler policy to pick a chunk count of 1, which disables parallelism entirely!

# Solution
Let's just set the minimum_chink_size to 1 for most map reduce operations.

Reviewed By: samwgoldman

Differential Revision: D58282862

fbshipit-source-id: 1ea85bdf80b177bc77040fbfd479f4b3e82c0f3c