Clean up from DataclassOptions.ml refactor
Summary:
When we created `dataclassOptions.ml` in order to make
the options visible to type check (which we did to address
conformance tests that require throwing type errors on certain
kinds of invalid dataclass definitions), we move the queries
over.

But then we decoupled the actual options extraction from the queries,
so that we could move them back to attributeResolution. This commit
makes that move. In the process I also eliminate the nested
`ExtractDataclassOptions` namespace, which isn't really needed once
we get `Queries` out of the way, and avoid `open DataclassOptions` in
downstream code (when possible I try to avoid global `open`, it makes
OCaml name resolution a little less obvious, especially if there are
a lot of big modules defining lots of symbols where collisions are
likely)

Reviewed By: migeed-z

Differential Revision: D58244121

fbshipit-source-id: 6c33f2c53165803b0686cb5604720d2214daa57f