Rename `apply`, add a comment describing it.
Summary:
The name `apply` is an artifact from when this was nested inside
a module; now that it isn't nested, the name carries no information
about what's actually happening, so we should make it more descriptive.

I also added a comment with a little more detail than I can put in the
name itself, in particular emphasizing the use of mutation.

Reviewed By: yangdanny97

Differential Revision: D58242819

fbshipit-source-id: 6d3f5f8dfd3c88b25876b21bbe0c50e5eb76463c