Make AnnotatedFunctionGenerator serializable [1/n multiprocessing]
Summary:
# Context
We want to speed up the python model generation step in Pysa.
We can see that the CPU usage is pretty low during model generation.

# Solution
We should use python multiprocessing to speed up the model generation.

# This diff
First, we need to make `AnnotatedFunctionGenerator` serializable so it can be used with the multiprocessing module. For that, we should provide exclude patterns as strings instead of compiled regular expressions.

Reviewed By: tianhan0

Differential Revision: D58290304

fbshipit-source-id: bc0d11ed5ae3db1d3c569da53ffd8631bd7d5d45