Move preprocess_alias_value to type.ml
Summary:
Context:
We want to handle type variables in type.ml instead of typeAliasEnvironement.ml in preparation to support PEP695. This is because we will no longer be able to detect type valiables via type aliasing due to the syntax change. As a result, we will incrementally move the code which handles type variables away from typeAliasEnvironement and eventually, we want to change the way we hangle type variables entirely.

Changes:
preprocess_alias_value needs to be in type.ml as part of this refactor, so we move it.

Reviewed By: stroxler

Differential Revision: D58267126

fbshipit-source-id: 20335f9655b1f3faf7c9ecb62e0e23a499d94518