Refactor typeCheckTest a bit to prepare for splitting
Summary:
I have been frequently running into test failures in `typeCheckTest.ml`
when I make changes to Pyre, and they are nearly impossible to understand without
hacking up the code, because the tests have hundreds of cases and we don't usually
get clear line numbers or indications of which test failed.

So I want to split the tests and use labeled test cases with line numbers. But
the `typeCheckTest.ml` code is a little hard to split because it is often
relying on the ability to mix together `let` bindings and unit-valued
computations. Since test splitting itself tends to be a big diff that I want to
be mostly mechanical, I decided to put this smallish refactor into its own diff;
this commit breaks up and/or rearranges some of the tests so that all the
let-bindings used by test cases can live at the top of the test case sequence, which
will make it much easier to use `test_list` when splitting cases.

Reviewed By: yangdanny97

Differential Revision: D58374892

fbshipit-source-id: 9748d4caf7ae7c6b0788302e01c805f44ebd6282