Simplify refinement logic: no upcasts
Summary:
I was looking at callers of `GlobalResolution.type_less_or_equal_either_way`
because it had some impact on an `typeInfo.ml` change I wanted to make, and
was a little surprised to see that refinement appears to be checking for
less-or-equal in *either* direction when refining.

This implies that we're willing to *upcast* a variable on an isinstance check,
which seems pretty iffy to me, `isinstance` is supposed to narrow types but not
widen them

(We might ask why someone would use isinstance if they didn't want
a cast... remember it has runtime behavior too, they might actually just need
it for business logic, or to validate a type that is "known" statically but
only in a gradual way that can't be fully trusted). If the actually want a
cast, that's what `cast` is for.)

I figured I'd try this change and see what the result is. In the unit tests,
the impact is minimal... one of the refinement unit tests cases changes,
we don't have a single integration test that relies on this, and a couple
of `pyre infer` behaviors change; I actually think the existing `pyre infer`
behaviors were pretty questionable so I'm happy with this change.

Reviewed By: grievejia

Differential Revision: D58375764

fbshipit-source-id: 19dd24435f4dfaf774222f89624934a40232ea9e