Parse using multiple partial sink kinds in the same rule
Summary:
The syntax of multi-source rules is designed to be the same as other taint
analysis tools, so that the user experience is the same across tools.

**Combined source rules.** Now we have a section called `rules` that contain
two sub-rules. This syntax is similar to other tools. Note that the syntax no
longer supports specifying main / secondary flows, because such support is
already deprecated by the backend. We can bring this back if needed (which
needs further investigation).

**String combine rules.** They now look exactly the same as combined source
rules, which reflects the fact the they share the same underlying
implementation.  As a result, we now require users to specify both partial sink
kinds in a rule. By contrast, we previously only ask a user to specify a single
partial sink kind, and auto-generate the kinds (with labels).

An alternative design is to keep the existing behavior -- specifying a single
partial sink kind and then auto-generating different kinds for the two
sub-rules. But I think it is simpler for both the users and the implementation
if string combine rules are syntactically similar to combined source rules.

In addition, the syntax also no longer supports specifying main / secondary
flows.

**Implementation.** The main changes are:
1. Module `CombinedSourceRuleSources` is renamed and updated, because we no
longer have a concept of "labels".
2. Tests in `configurationTest.ml`.

Note that the old syntax are still supported. Otherwise landing this diff would
break the CI tests (which run against `master`), which is not ideal.  To test
this support, I kept some of the old tests that parse the old syntax. I
deliberately dropped the tests that test against parsing errors, to simplify
the diff. The old tests will be removed once the rules in production are
migrated.

Reviewed By: arthaud

Differential Revision: D58260058

fbshipit-source-id: fcbeb821b87f94c42e0d37b1ef63190d885bcc43